split build.rs
